add scrcpy.sh for build for linux (include arm64)

+++ win-git/scrcpy.sh
@@ -11,35 +11,59 @@ SOFT_ARCH=64
 
 # uname Linux .bashrc uname Darwin MINGW64 .bash_profile
 PLATFORM=$(platform)
-case $(arch) in 
-  amd64) SOFT_ARCH=64;;
-  386) SOFT_ARCH=32;;
-esac
-
-SOFT_FILE_NAME=${NAME}-${PLATFORM}${SOFT_ARCH}-${SOFT_VERSION}
-SOFT_FILE_PACK=$(soft_file_pack $SOFT_FILE_NAME)
-# init pwd
-cd $HOME
-
-SOFT_URL=https://github.com/Genymobile/${NAME}/releases/download/${SOFT_VERSION}/${SOFT_FILE_PACK}
-#if [[ "$(${NAME} --version)" != *${NAME}\ ${SOFT_VERSION}* ]]; then
-if [[ "$(${NAME} --version)" != *${NAME}\ ${SOFT_VERSION}* ]]; then
-  $(cache_downloader $SOFT_FILE_PACK $SOFT_URL)
-  $(cache_unpacker $SOFT_FILE_PACK $SOFT_FILE_NAME)
-  
-  rm -rf ${SOFT_HOME} && \
-    mv $(cache_folder)/${SOFT_FILE_NAME} ${SOFT_HOME} 
+if [[ $(platform) == *linux* ]]; then
+  #deb apt
+  sudo apt install ffmpeg libsdl2-2.0-0 gcc git pkg-config meson ninja-build \
+                 libavcodec-dev libavformat-dev libavutil-dev \
+                 libsdl2-dev openjdk-11-jdk -y
+  cd $(install_path)
+  git clone https://github.com/Genymobile/${NAME} $SOFT_HOME
+  cd $SOFT_HOME
+  git pull
+  #prebuilt server
+  SOFT_URL=https://github.com/Genymobile/${NAME}/releases/download/${SOFT_VERSION}/${NAME}-server-${SOFT_VERSION}
+  SOFT_FILE_NAME=${NAME}-server-${SOFT_VERSION}
+  $(cache_downloader $SOFT_FILE_NAME $SOFT_URL)
+  mv $(cache_folder)/${SOFT_FILE_NAME} ${SOFT_HOME}/
+  #build
+  meson x --buildtype release --strip -Db_lto=true \
+      -Dprebuilt_server=${SOFT_FILE_NAME}
+  ninja -Cx
+  #install
+  #sudo ninja -Cx install
+  echo 'alias scrcpy='"${SOFT_HOME}/run ${SOFT_HOME}/x">>${TOOLSRC}
 fi
-#--------------new .toolsrc-----------------------
-export PATH=$PATH:${SOFT_HOME}
+if [[ $(platform) == *win* ]]; then
+  case $(arch) in 
+    amd64) SOFT_ARCH=64;;
+    386) SOFT_ARCH=32;;
+  esac
 
-echo 'export PATH=$PATH:'${SOFT_HOME}>>${TOOLSRC}
+  SOFT_FILE_NAME=${NAME}-${PLATFORM}${SOFT_ARCH}-${SOFT_VERSION}
+  SOFT_FILE_PACK=$(soft_file_pack $SOFT_FILE_NAME)
+  # init pwd
+  cd $HOME
 
-#  ----windows bat----
-if [[ $WIN_PATH  ]]; then
-  if [[ $PLATFORM == windows ]]; then
-    windowsENV="$(echo -e ${PATH//:/;\\n}';' |sort|uniq|cygpath -w -f -|tr -d '\n')"
-    echo $windowsENV
-    setx Path "$windowsENV"
+  SOFT_URL=https://github.com/Genymobile/${NAME}/releases/download/${SOFT_VERSION}/${SOFT_FILE_PACK}
+  #if [[ "$(${NAME} --version)" != *${NAME}\ ${SOFT_VERSION}* ]]; then
+  if [[ "$(${NAME} --version)" != *${NAME}\ ${SOFT_VERSION}* ]]; then
+    $(cache_downloader $SOFT_FILE_PACK $SOFT_URL)
+    $(cache_unpacker $SOFT_FILE_PACK $SOFT_FILE_NAME)
+    
+    rm -rf ${SOFT_HOME} && \
+      mv $(cache_folder)/${SOFT_FILE_NAME} ${SOFT_HOME} 
+  fi
+  #--------------new .toolsrc-----------------------
+  export PATH=$PATH:${SOFT_HOME}
+
+  echo 'export PATH=$PATH:'${SOFT_HOME}>>${TOOLSRC}
+
+  #  ----windows bat----
+  if [[ $WIN_PATH  ]]; then
+    if [[ $PLATFORM == windows ]]; then
+      windowsENV="$(echo -e ${PATH//:/;\\n}';' |sort|uniq|cygpath -w -f -|tr -d '\n')"
+      echo $windowsENV
+      setx Path "$windowsENV"
+    fi
   fi
 fi
